Fixed small bug and compilation warning in fibonacci.c
Removed recursion inside menu() function. That was unnecessary, and caused the program to keep running when it should terminate, in case a user typed other char than 'y' or 'n' when prompted, and then tried to exit the program by typing 'y' (the program would run once again instead of terminating). This commit also fixes the compilation warning: non-void function does not return a value in all control paths.